experiments/requirements.txt
colorama==0.4.6
numpy==1.25.2
paho-mqtt==1.6.1
pandas==2.0.3
python-dateutil==2.8.2
pytz==2023.3
tqdm==4.66.1
tzdata==2023.3